add pug templates support

+++ .eslintrc.js
@@ -10,7 +10,7 @@ module.exports = {
   rules: {
     indent: ["error", 2],
     "linebreak-style": ["error", "unix"],
-    quotes: ["error", "single"],
+    quotes: ["error", "double"],
     semi: ["error", "always"]
   }
 };
